Split PortAdmin SNMP handlers into submodule tree
Because:
- The entire management back-end of PortAdmin needs to be refactored to
  accommodate multiple protocols and backends in the near future.
- Splitting each vendor-specific handler into separate files also makes
  sense from a code readability standpoint, IMHO.
